feat!: y-axis refactor start
