+++ src/chart/components/y_axis/y-axis-global.component.ts
@@ -1,0 +1,60 @@
+import { BarType, FullChartColors } from '../../chart.config';
+import { PriceMovement } from '../../model/candle-series.model';
+import { ChartBaseElement } from '../../model/chart-base-element';
+import { DataSeriesType } from '../../model/data-series.config';
+import { resolveColorForArea, resolveColorForBar, resolveColorForBaseLine, resolveColorForCandle, resolveColorForHistogram, resolveColorForLine, resolveColorForScatterPlot, resolveColorForTrendAndHollow, resolveDefaultColorForLabel } from './label-color.functions';
+
+export type LabelColorResolver = (priceMovement: PriceMovement, colors: FullChartColors) => string;
+
+export class YAxisGlobalComponent extends ChartBaseElement {
+	private labelsColorByChartTypeMap: Partial<Record<DataSeriesType, LabelColorResolver>> = {};
+
+    // public yAxisLabelsModel: YAxisLabelsModel;
+
+	constructor() {
+        super();
+        // this.yAxisLabelsModel = new YAxisLabelsModel();
+        // this.addChildEntity(this.yAxisLabelsModel);
+		this.registerDefaultLabelColorResolver();
+	}
+
+	/**
+	 * Registers default label color resolvers for different chart types.
+	 * @private
+	 * @function
+	 * @name registerDefaultLabelColorResolver
+	 * @returns {void}
+	 */
+	private registerDefaultLabelColorResolver() {
+		this.registerLabelColorResolver('candle', resolveColorForCandle);
+		this.registerLabelColorResolver('bar', resolveColorForBar);
+		this.registerLabelColorResolver('line', resolveColorForLine);
+		this.registerLabelColorResolver('area', resolveColorForArea);
+		this.registerLabelColorResolver('scatterPlot', resolveColorForScatterPlot);
+		this.registerLabelColorResolver('histogram', resolveColorForHistogram);
+		this.registerLabelColorResolver('baseline', resolveColorForBaseLine);
+		this.registerLabelColorResolver('trend', resolveColorForTrendAndHollow);
+		this.registerLabelColorResolver('hollow', resolveColorForTrendAndHollow);
+	}
+
+	/**
+	 * Registers a label color resolver for a specific chart type.
+	 *
+	 * @param {BarType} chartType - The type of chart for which the label color resolver is being registered.
+	 * @param {LabelColorResolver} resolver - The function that will be used to resolve the color of the labels for the specified chart type.
+	 * @returns {void}
+	 */
+	public registerLabelColorResolver(chartType: BarType, resolver: LabelColorResolver) {
+		this.labelsColorByChartTypeMap[chartType] = resolver;
+	}
+
+	/**
+	 * Returns a function that resolves the color for a label based on the type of data series.
+	 * @param {DataSeriesType} candlesType - The type of data series.
+	 * @returns {Function} - A function that resolves the color for a label.
+	 * If there is no color mapping for the given data series type, it returns the default color resolver function.
+	 */
+	public getLabelsColorResolver(candlesType: DataSeriesType) {
+		return this.labelsColorByChartTypeMap[candlesType] ?? resolveDefaultColorForLabel;
+	}
+}
