+++ src/chart/components/y_axis/y-axis.component.ts
@@ -6,51 +6,33 @@
 import { Subject } from 'rxjs';
 import { CanvasBoundsContainer, CanvasElement, CHART_UUID } from '../../canvas/canvas-bounds-container';
 import { CursorHandler } from '../../canvas/cursor.handler';
+import { YAxisWidthContributor } from '../../canvas/y-axis-bounds.container';
 import {
-	BarType,
 	ChartConfigComponentsYAxis,
-	FullChartColors,
 	FullChartConfig,
 	YAxisAlign,
 	YAxisLabelAppearanceType,
 	YAxisLabelMode,
-	YAxisLabelType,
+	YAxisLabelType
 } from '../../chart.config';
 import { ClearCanvasDrawer } from '../../drawers/clear-canvas.drawer';
 import { CompositeDrawer } from '../../drawers/composite.drawer';
 import { DrawingManager } from '../../drawers/drawing-manager';
 import EventBus from '../../events/event-bus';
 import { CanvasInputListenerComponent } from '../../inputlisteners/canvas-input-listener.component';
-import { PriceMovement } from '../../model/candle-series.model';
 import { CanvasModel } from '../../model/canvas.model';
 import { ChartBaseElement } from '../../model/chart-base-element';
-import { DataSeriesType } from '../../model/data-series.config';
 import { ScaleModel } from '../../model/scale.model';
 import { uuid } from '../../utils/uuid.utils';
 import { ChartModel } from '../chart/chart.model';
 import { PriceAxisType } from '../labels_generator/numeric-axis-labels.generator';
 import { ChartPanComponent } from '../pan/chart-pan.component';
 import { PaneManager } from '../pane/pane-manager.component';
-import {
-	resolveColorForArea,
-	resolveColorForBar,
-	resolveColorForBaseLine,
-	resolveColorForCandle,
-	resolveColorForHistogram,
-	resolveColorForLine,
-	resolveColorForScatterPlot,
-	resolveColorForTrendAndHollow,
-	resolveDefaultColorForLabel,
-} from './label-color.functions';
-import { LastCandleLabelsProvider } from './price_labels/last-candle-labels.provider';
 import { LabelsGroups, VisualYAxisLabel, YAxisLabelsProvider } from './price_labels/y-axis-labels.model';
 import { YAxisPriceLabelsDrawer } from './price_labels/y-axis-price-labels.drawer';
 import { YAxisScaleHandler } from './y-axis-scale.handler';
 import { YAxisDrawer } from './y-axis.drawer';
 import { YAxisModel } from './y-axis.model';
-import { YAxisWidthContributor } from '../../canvas/y-axis-bounds.container';
-
-export type LabelColorResolver = (priceMovement: PriceMovement, colors: FullChartColors) => string;
 
 /**
  * Y axis component. Contains all Y axis related logic.
@@ -60,7 +42,6 @@ export class YAxisComponent extends ChartBaseElement {
 	yAxisModel: YAxisModel;
 	public axisTypeSetSubject: Subject<PriceAxisType> = new Subject<PriceAxisType>();
 	public readonly state: ChartConfigComponentsYAxis;
-	private labelsColorByChartTypeMap: Partial<Record<DataSeriesType, LabelColorResolver>> = {};
 	public drawer: CompositeDrawer;
 
 	constructor(
@@ -69,7 +50,7 @@ export class YAxisComponent extends ChartBaseElement {
 		private canvasModel: CanvasModel,
 		private yAxisLabelsCanvasModel: CanvasModel,
 		private backgroundCanvasModel: CanvasModel,
-		private chartModel: ChartModel,
+		chartModel: ChartModel,
 		private scaleModel: ScaleModel,
 		canvasInputListeners: CanvasInputListenerComponent,
 		private canvasBoundsContainer: CanvasBoundsContainer,
@@ -77,6 +58,8 @@ export class YAxisComponent extends ChartBaseElement {
 		chartPanComponent: ChartPanComponent,
 		paneManager: PaneManager,
 		private cursorHandler: CursorHandler,
+		private paneUUID: string,
+		private extentIdx: number,
 	) {
 		super();
 		this.state = config.components.yAxis;
@@ -84,7 +67,6 @@ export class YAxisComponent extends ChartBaseElement {
 		this.drawer = yAxisCompositeDrawer;
 		const clearYAxis = new ClearCanvasDrawer(this.yAxisLabelsCanvasModel);
 		yAxisCompositeDrawer.addDrawer(clearYAxis, 'YAXIS_CLEAR');
-		this.registerDefaultLabelColorResolver();
 
 		drawingManager.addDrawer(yAxisCompositeDrawer, 'Y_AXIS');
 
@@ -103,7 +85,7 @@ export class YAxisComponent extends ChartBaseElement {
 		//#endregion
 
 		this.yAxisModel = new YAxisModel(
-			paneManager.paneComponents[CHART_UUID],
+			paneManager.paneComponents[this.paneUUID],
 			eventBus,
 			this.config,
 			canvasBoundsContainer,
@@ -125,7 +107,7 @@ export class YAxisComponent extends ChartBaseElement {
 			() => this.yAxisModel.yAxisBaseLabelsModel.labels,
 			() => canvasBoundsContainer.getBounds(CanvasElement.Y_AXIS),
 			() => this.state.visible,
-			chartModel.pane.scaleModel.toY.bind(chartModel.pane.scaleModel),
+			scaleModel.toY.bind(scaleModel),
 		);
 		yAxisCompositeDrawer.addDrawer(yAxisDrawer);
 		//#endregion
@@ -143,14 +125,14 @@ export class YAxisComponent extends ChartBaseElement {
 
 		//#region init YAxisLabels related stuff
 		// default labels provider
-		const lastCandleLabelsProvider = new LastCandleLabelsProvider(
-			this.chartModel,
-			this.config,
-			this.chartModel.lastCandleLabelsByChartType,
-			this.getLabelsColorResolver.bind(this),
-		);
+		// const lastCandleLabelsProvider = new LastCandleLabelsProvider(
+		// 	this.chartModel,
+		// 	this.config,
+		// 	this.chartModel.lastCandleLabelsByChartType,
+		// 	this.getLabelsColorResolver.bind(this),
+		// );
 
-		this.registerYAxisLabelsProvider(lastCandleLabelsProvider, LabelsGroups.MAIN);
+		// this.registerYAxisLabelsProvider(lastCandleLabelsProvider, LabelsGroups.MAIN);
 		//#endregion
 
 		// TODO hack, remove when each pane will have separate y-axis component
@@ -161,12 +143,12 @@ export class YAxisComponent extends ChartBaseElement {
 	private updateCursor() {
 		if (this.state.type === 'percent') {
 			this.cursorHandler.setCursorForCanvasEl(
-				CanvasElement.PANE_UUID_Y_AXIS(CHART_UUID),
+				CanvasElement.PANE_UUID_Y_AXIS(this.paneUUID, this.extentIdx),
 				this.config.components.yAxis.resizeDisabledCursor,
 			);
 		} else {
 			this.cursorHandler.setCursorForCanvasEl(
-				CanvasElement.PANE_UUID_Y_AXIS(CHART_UUID),
+				CanvasElement.PANE_UUID_Y_AXIS(this.paneUUID, this.extentIdx),
 				this.config.components.yAxis.cursor,
 			);
 		}
@@ -177,34 +159,6 @@ export class YAxisComponent extends ChartBaseElement {
 	 */
 	public updateOrderedLabels(adjustYAxisWidth = false) {
 		this.yAxisModel.yAxisLabelsModel.updateLabels(adjustYAxisWidth);
-	}
-
-	/**
-	 * Calls the parent class's doActivate method
-	 * @protected
-	 * @returns {void}
-	 */
-	protected doActivate(): void {
-		super.doActivate();
-	}
-
-	/**
-	 * Registers default label color resolvers for different chart types.
-	 * @private
-	 * @function
-	 * @name registerDefaultLabelColorResolver
-	 * @returns {void}
-	 */
-	private registerDefaultLabelColorResolver() {
-		this.registerLabelColorResolver('candle', resolveColorForCandle);
-		this.registerLabelColorResolver('bar', resolveColorForBar);
-		this.registerLabelColorResolver('line', resolveColorForLine);
-		this.registerLabelColorResolver('area', resolveColorForArea);
-		this.registerLabelColorResolver('scatterPlot', resolveColorForScatterPlot);
-		this.registerLabelColorResolver('histogram', resolveColorForHistogram);
-		this.registerLabelColorResolver('baseline', resolveColorForBaseLine);
-		this.registerLabelColorResolver('trend', resolveColorForTrendAndHollow);
-		this.registerLabelColorResolver('hollow', resolveColorForTrendAndHollow);
 	}
 
 	//#region public methods
@@ -346,27 +300,6 @@ export class YAxisComponent extends ChartBaseElement {
 		this.state.labels.descriptions = descVisibility;
 		//  recalculating labels is not needed, so just redraw YAxis
 		this.canvasModel.fireDraw();
-	}
-
-	/**
-	 * Registers a label color resolver for a specific chart type.
-	 *
-	 * @param {BarType} chartType - The type of chart for which the label color resolver is being registered.
-	 * @param {LabelColorResolver} resolver - The function that will be used to resolve the color of the labels for the specified chart type.
-	 * @returns {void}
-	 */
-	public registerLabelColorResolver(chartType: BarType, resolver: LabelColorResolver) {
-		this.labelsColorByChartTypeMap[chartType] = resolver;
-	}
-
-	/**
-	 * Returns a function that resolves the color for a label based on the type of data series.
-	 * @param {DataSeriesType} candlesType - The type of data series.
-	 * @returns {Function} - A function that resolves the color for a label.
-	 * If there is no color mapping for the given data series type, it returns the default color resolver function.
-	 */
-	public getLabelsColorResolver(candlesType: DataSeriesType) {
-		return this.labelsColorByChartTypeMap[candlesType] ?? resolveDefaultColorForLabel;
 	}
 	//#endregion
 }
